Move helper method findLargestMatchingBreakpoint
